+++ settings.php
@@ -29,16 +29,33 @@ defined('MOODLE_INTERNAL') || die;
 if ($ADMIN->fulltree) {
 
     $defaulthost = 'https://documentserver.url';
+    $linktodocs = 'https://www.onlyoffice.com/docs-registration.aspx?referer=moodle';
     $defaultjwtheader = 'Authorization';
+    $bannerdata = [
+        'title' => get_string('banner_title', 'onlyofficeeditor'),
+        'description' => get_string('banner_description', 'onlyofficeeditor'),
+        'link' => [
+            'title' => get_string('banner_link_title', 'onlyofficeeditor'),
+            'href' => $linktodocs,
+        ],
+    ];
+    $banner = $OUTPUT->render_from_template('mod_onlyofficeeditor/banner', $bannerdata);
     $settings->add(new admin_setting_configtext('onlyofficeeditor/documentserverurl',
         get_string('documentserverurl', 'onlyofficeeditor'), get_string('documentserverurl_desc', 'onlyofficeeditor'),
         $defaulthost));
+    $settings->add(new admin_setting_configcheckbox('onlyofficeeditor/disable_verify_ssl',
+        get_string('disable_verify_ssl', 'onlyofficeeditor'), '', 0));
     $settings->add(new admin_setting_configtext('onlyofficeeditor/documentserversecret',
         get_string('documentserversecret', 'onlyofficeeditor'), get_string('documentserversecret_desc', 'onlyofficeeditor'), ''));
     $settings->add(new admin_setting_configtext('onlyofficeeditor/jwtheader',
         get_string('jwtheader', 'onlyofficeeditor'), '', $defaultjwtheader));
     $settings->add(new admin_setting_configcheckbox('onlyofficeeditor/forcesave',
         get_string('forcesave', 'onlyofficeeditor'), '', 0));
+    $settings->add(new admin_setting_heading('onlyofficeeditor/banner', '', $banner));
+    $settings->add(new admin_setting_configtext('onlyofficeeditor/documentserverinternal',
+        get_string('documentserverinternal', 'onlyofficeeditor'), '', ''));
+    $settings->add(new admin_setting_configtext('onlyofficeeditor/storageurl',
+        get_string('storageurl', 'onlyofficeeditor'), '', ''));
     $settings->add(new admin_setting_heading('onlyofficeeditor/editor_view',
         get_string('editor_view', 'onlyofficeeditor'), ''));
     $settings->add(new admin_setting_configcheckbox('onlyofficeeditor/editor_view_chat',